requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
pandas>=2.0.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.13
python-dotenv>=1.0
fastapi>=0.110
uvicorn>=0.29
httpx>=0.27